+++ FilmsAboutBack/Services/Interfaces/IRatingService.cs
@@ -9,6 +9,6 @@ namespace FilmsAboutBack.Services.Interfaces
     public interface IRatingService : ICRUDService<Rating>
     {
         Task<int> GetByPairIdAsync(int userId, int filmId);
-        Task<float> GetRatingByIdAsync(int filmId);
+        Task<double> GetRatingByIdAsync(int filmId);
     }
 }
